created bunch of .c files by category and moved realted functions, fixes issue #3
