+++ .gitignore
@@ -3,3 +3,4 @@ bin/**
 perdita
 version.h
 *.zip
+obj/
