+++ headers/psv.h
@@ -1,0 +1,12 @@
+#include <stdio.h>
+#include <stddef.h>
+
+#include "globals.h"
+
+#ifndef PSV_H
+#define PSV_H
+
+    void open_psv_raw_file(void);
+    void close_psv_raw_file(void);
+
+#endif
